+++ test/StorageBase/Bridge/Psr11StorageFacadeTest.php
@@ -283,13 +283,6 @@ class Psr11StorageFacadeTest extends TestCase
             static::assertSame('Action not found', $throwable->getPrevious()->getMessage());
         }
 
-        try {
-            $facade->getRouteCapabilityOverviewAction();
-            static::fail();
-        } catch (StorageFacadeServiceExceptionInterface $throwable) {
-            static::assertSame('Action not found', $throwable->getPrevious()->getMessage());
-        }
-
         static::assertTrue(true, 'We just do not expect an exception');
     }
 
@@ -337,7 +330,6 @@ class Psr11StorageFacadeTest extends TestCase
         $facade->getPortalNodeStorageListAction();
         $facade->getPortalNodeStorageSetAction();
         $facade->getStorageKeySerializer();
-        $facade->getRouteCapabilityOverviewAction();
 
         static::assertTrue(true, 'We just do not expect an exception');
     }