+++ test/Utility/TaggedDateCollectionTest.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Utility\Test;
 
 use Heptacom\HeptaConnect\Utility\Collection\AbstractCollection;
+use Heptacom\HeptaConnect\Utility\Collection\AbstractIterable;
 use Heptacom\HeptaConnect\Utility\Collection\Contract\AbstractTaggedCollection;
 use Heptacom\HeptaConnect\Utility\Collection\Contract\TagItem;
+use Heptacom\HeptaConnect\Utility\Collection\IterableImplementationTrait;
 use Heptacom\HeptaConnect\Utility\Date\Date;
 use Heptacom\HeptaConnect\Utility\Date\DateCollection;
 use Heptacom\HeptaConnect\Utility\Date\TaggedDateCollection;
@@ -15,7 +17,9 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 #[CoversClass(AbstractCollection::class)]
+#[CoversClass(AbstractIterable::class)]
 #[CoversClass(AbstractTaggedCollection::class)]
+#[CoversClass(IterableImplementationTrait::class)]
 #[CoversClass(TagItem::class)]
 #[CoversClass(DateCollection::class)]
 #[CoversClass(TaggedDateCollection::class)]